Alexandre/escolha tipo de usuario

+++ projeto-empresarial/routes/web.php
@@ -17,11 +17,11 @@ use App\Http\Controllers\UserController;
 */
 
 Route::get('/', [UserController::class, 'index'])->name('index');
-Route::get('/user/create', [UserController::class, 'create'])->name('user.create');
-Route::post('/user/create', [RegisteredUserController::class, 'store'])->name('user.store');
+Route::get('/user/create', [UserController::class, 'create'])->middleware(['auth', 'is_admin'])->name('user.create');
+Route::post('/user/create', [RegisteredUserController::class, 'store'])->middleware(['auth', 'is_admin'])->name('user.store');
 Route::put('/edit/{id}', [UserController::class, "update"])->name("user.update");
 Route::get('/edit/{id}', [UserController::class, 'edit'])->name('user.edit');
 Route::get('/dashboard', [UserController::class, 'dashboard'])->middleware(['auth', 'is_admin'])->name('dashboard');
-Route::delete('/delete/{id}', [UserController::class, 'delete'])->name('user.delete');
+Route::delete('/delete/{id}', [UserController::class, 'delete'])->middleware(['auth', 'is_admin'])->name('user.delete');
 
 require __DIR__ . '/auth.php';
